requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
python-multipart==0.0.18
pydantic[email]==2.10.4
httpx==0.28.1
PyPDF2==3.0.1
chromadb==0.5.23
jinja2==3.1.5
slowapi==0.1.9
PyMuPDF
python-pptx
python-docx
cachetools
psycopg2-binary
